Move PrimaryNavMenu to layout directory
