+++ src/view/layout/PrimaryNavMenu.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { NavLink } from "react-router-dom";
-import UsaButton from "./util/UsaButton";
+import UsaButton from "../util/UsaButton";
 import close from "uswds/dist/img/close.svg";
 import "./PrimaryNavMenu.css";
